# docs/spec/openapi.yaml
openapi: 3.1.0

info:
  title: API de Localização
  version: 1.0.0
tags:
  - name: Localização

paths:
  /cities:
    get:
      summary: Buscar cidades
      operationId: cities/search
      tags:
        - Localização
      parameters:
        - name: query
          in: query
          description: O nome da cidade a ser buscada
          required: true
          schema:
            type: string
          example: São Paulo
      responses:
        '200':
          description: Cidades encontradas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/City'
        '400':
          description: Erro de validação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '5XX':
          description: Erro inesperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'

  /cities/distances:
    get:
      summary: Obter a distância entre duas cidades
      operationId: cities/distances/get
      tags:
        - Localização
      parameters:
        - name: originCityId
          in: query
          description: O identificador da cidade de origem
          required: true
          schema:
            type: string
          example: aGVyZTpjbTpuYW1lZHBsYWNlOjIzMDMwNjEy
        - name: destinationCityId
          in: query
          description: O identificador da cidade de destino
          required: true
          schema:
            type: string
          example: aGVyZTpjbTpuYW1lZHBsYWNlOjIzMDM2NjI4
      responses:
        '200':
          description: Distância calculada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Distance'
        '400':
          description: Erro de validação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '404':
          description: Uma ou ambas as cidades não encontradas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '5XX':
          description: Erro inesperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'

components:
  schemas:
    City:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          description: O identificador da cidade
          example: aGVyZTpjbTpuYW1lZHBsYWNlOjIzMDMwNjEy
        name:
          type: string
          description: O nome da cidade
          example: São Paulo
        stateName:
          type: string
          description: O nome do estado
          example: São Paulo
        stateCode:
          type: string
          description: O código do estado
          example: SP
        countryName:
          type: string
          description: O nome do país
          example: Brasil
        countryCode:
          type: string
          description: O código do país
          example: BRA

    Distance:
      type: object
      required:
        - kilometers
      properties:
        kilometers:
          type: number
          description: A distância em quilômetros (km)
          format: double
          example: 145.3

    ValidationError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: A mensagem de erro
          example: Validation error
        issues:
          type: array
          description: Os problemas de validação
          items:
            type: object
            properties:
              message:
                type: string
                description: A mensagem de erro
                example: 'Invalid input: expected string, received number'
              code:
                type: string
                description: O código do erro
                example: invalid_type
              path:
                type: array
                description: O caminho do erro
                items:
                  oneOf:
                    - type: string
                    - type: number
                example: ['names', 1]
            additionalProperties: true

    NotFoundError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: A mensagem de erro
          example: Not found

    InternalServerError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: A mensagem de erro
          example: Internal server error